requirements.txt
torch
Pillow
moviepy
opencv-python
tkinter
pygame